Moved helper function to utils dir, disabled sort.
