+++ src/routes/projects/projects.py
@@ -1,35 +1,13 @@
-from flask import Blueprint, jsonify
+from flask import Blueprint, current_app, jsonify
 from dotenv import load_dotenv
+from src.utils.project_img import ParseOSCrepo
 import requests
-from html.parser import HTMLParser
 
 load_dotenv()
 projects_bp = Blueprint("projects_bp", __name__, url_prefix="/projects")
 
 
-class ParseOSCrepo(HTMLParser):  # The HTML parser for scraping img link from
-    token: str = None
-
-    def handle_starttag(self, tag: str, attrs: str):
-        if self.token:
-            return
-        if tag != "meta":
-            return
-        token = None
-        for (index, (i, j)) in enumerate(attrs):
-            if i == "content":
-                token = j
-            if all([i == "property", j == "og:image"]):
-                if token:
-                    self.token = token
-                    return
-                for (inner_index, (ni, nj)) in enumerate(attrs, start=index):
-                    if ni == "content":
-                        self.token = nj
-                        return
-
-
-def imglink(github_repolink):  # Web scraping the image link from github
+def img_link(github_repolink):  # Web scraping the image link from github
     req = requests.get(f"{github_repolink}")
     parseObj = ParseOSCrepo()
     parseObj.feed(req.text)
@@ -38,6 +16,7 @@ def imglink(github_repolink):  # Web scraping the image link from github
 
 @projects_bp.route("/", methods=["GET"])
 def project_info():
+    current_app.config["JSON_SORT_KEYS"] = False
     results = []
     req = requests.get(
         "https://api.github.com/users/Open-Source-Community-VIT-AP/repos"
@@ -50,7 +29,7 @@ def project_info():
                 "Stars": i["stargazers_count"],
                 "Name": i["name"],
                 "Description": i["description"],
-                "Image": imglink(i["html_url"]),
+                "Image": img_link(i["html_url"]),
                 "Repository_link": i["html_url"],
                 "SSH": i["ssh_url"],
             }
@@ -58,10 +37,10 @@ def project_info():
 
     for i in range(len(results)):  # Sorting according to stars
         for j in range(len(results) - 1):
-            if results[j]["stars"] < results[j + 1]["stars"]:
+            if results[j]["Stars"] < results[j + 1]["Stars"]:
                 results[j], results[j + 1] = results[j + 1], results[j]
 
     for i in results:  # Removing stars from the results
-        del i["stars"]
+        del i["Stars"]
 
     return jsonify(results[0:10])
